Add default port. Move handlers. Update serving message.
