+++ demo/src/mockApiWorker.ts
@@ -1,5 +1,5 @@
 import { setupWorker } from "msw";
 
-import { handlers } from "./handlers";
+import { handlers } from "../../src/mocks/handlers";
 
 export const mockApiWorker = setupWorker(...handlers);
